+++ src/balances/balances.controller.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
-import safeBalanceFactory from '../datasources/transaction-service/entities/__tests__/balance.factory';
+import safeBalanceFactory from '../datasources/transaction-api/entities/__tests__/balance.factory';
 import exchangeResultFactory from '../datasources/exchange/entities/__tests__/exchange.factory';
 import chainFactory from '../datasources/config-service/entities/__tests__/chain.factory';
 import {
