+++ src/balances/balances.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
 import { BalancesService } from './balances.service';
 import { BalancesController } from './balances.controller';
-import { TransactionServiceModule } from '../datasources/transaction-service/transaction-service.module';
+import { TransactionApiModule } from '../datasources/transaction-api/transaction-api.module';
 import { ConfigServiceModule } from '../datasources/config-service/config-service.module';
 import { ExchangeModule } from '../datasources/exchange/exchange.module';
 
 @Module({
   controllers: [BalancesController],
   providers: [BalancesService],
-  imports: [TransactionServiceModule, ExchangeModule, ConfigServiceModule],
+  imports: [TransactionApiModule, ExchangeModule, ConfigServiceModule],
 })
 export class BalancesModule {}
