+++ src/datasources/transaction-api/transaction-api.manager.ts
@@ -2,16 +2,16 @@ import { Inject, Injectable, Logger } from '@nestjs/common';
 import { HttpErrorHandler } from '../errors/http-error-handler';
 import { Chain } from '../config-service/entities/chain.entity';
 import { ConfigService } from '../config-service/config-service.service';
-import { TransactionService } from './transaction-service.service';
+import { TransactionApi } from './transaction-api.service';
 import {
   INetworkService,
   NetworkService,
 } from '../../common/network/network.service.interface';
 
 @Injectable()
-export class TransactionServiceManager {
-  private readonly logger = new Logger(TransactionService.name);
-  private transactionServiceMap: Record<string, TransactionService> = {};
+export class TransactionApiManager {
+  private readonly logger = new Logger(TransactionApi.name);
+  private transactionApiMap: Record<string, TransactionApi> = {};
 
   constructor(
     private readonly configService: ConfigService,
@@ -19,20 +19,20 @@ export class TransactionServiceManager {
     private readonly httpErrorHandler: HttpErrorHandler,
   ) {}
 
-  async getTransactionService(chainId: string): Promise<TransactionService> {
+  async getTransactionService(chainId: string): Promise<TransactionApi> {
     this.logger.log(`Getting TransactionService instance for chain ${chainId}`);
-    const transactionService = this.transactionServiceMap[chainId];
+    const transactionService = this.transactionApiMap[chainId];
     if (transactionService !== undefined) return transactionService;
 
     this.logger.log(
       `Transaction Service for chain ${chainId} not available. Fetching from the Config Service`,
     );
     const chain: Chain = await this.configService.getChain(chainId);
-    this.transactionServiceMap[chainId] = new TransactionService(
+    this.transactionApiMap[chainId] = new TransactionApi(
       chain.transactionService,
       this.networkService,
       this.httpErrorHandler,
     );
-    return this.transactionServiceMap[chainId];
+    return this.transactionApiMap[chainId];
   }
 }
