+++ src/datasources/transaction-api/transaction-api.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { ConfigServiceModule } from '../config-service/config-service.module';
 import { HttpErrorHandler } from '../errors/http-error-handler';
-import { TransactionServiceManager } from './transaction-service.manager';
+import { TransactionApiManager } from './transaction-api.manager';
 
 @Module({
   imports: [ConfigServiceModule],
-  providers: [TransactionServiceManager, HttpErrorHandler],
-  exports: [TransactionServiceManager],
+  providers: [TransactionApiManager, HttpErrorHandler],
+  exports: [TransactionApiManager],
 })
-export class TransactionServiceModule {}
+export class TransactionApiModule {}
